tools/article-export/pyproject.toml
[project]
name = "article-export"
version = "0.1.0"
description = "Academic article export tooling for EduAssist Platform."
requires-python = ">=3.12"
dependencies = [
  "python-docx>=1.1.2",
  "PyYAML>=6.0.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/article_export"]
